+++ .idea/misc.xml
@@ -56,6 +56,7 @@
         <entry key="..\:/Users/cksgo/Desktop/Subject/UMC/dep_cocktail_dakk/cocktail_dakk_android/app/src/main/res/layout/fragment_search_tab_temp_result.xml" value="0.14990942028985507" />
         <entry key="..\:/Users/cksgo/Desktop/Subject/UMC/dep_cocktail_dakk/cocktail_dakk_android/app/src/main/res/layout/item_cocktail.xml" value="0.14990942028985507" />
         <entry key="..\:/Users/cksgo/Desktop/UMC_Android/backup/cocktaildakk_android/app/src/main/res/drawable/selector_main_filter.xml" value="0.21875" />
+        <entry key="..\:/Users/cksgo/Desktop/UMC_Android/backup/cocktaildakk_android/app/src/main/res/drawable/selector_navi_mypage.xml" value="0.21875" />
         <entry key="..\:/Users/cksgo/Desktop/UMC_Android/backup/cocktaildakk_android/app/src/main/res/drawable/setting_bg_btn_radius_bl_off.xml" value="0.21875" />
         <entry key="..\:/Users/cksgo/Desktop/UMC_Android/backup/cocktaildakk_android/app/src/main/res/drawable/setting_bg_btn_radius_bl_on.xml" value="0.21875" />
         <entry key="..\:/Users/cksgo/Desktop/UMC_Android/backup/cocktaildakk_android/app/src/main/res/drawable/setting_bg_btn_radius_bl_on_off.xml" value="0.21875" />
@@ -85,8 +86,9 @@
         <entry key="..\:/Users/cksgo/Desktop/UMC_Android/backup/cocktaildakk_android/app/src/main/res/layout/fragment_start_f_keyword.xml" value="0.3129528985507246" />
         <entry key="..\:/Users/cksgo/Desktop/UMC_Android/backup/cocktaildakk_android/app/src/main/res/layout/fragment_start_name.xml" value="0.27239583333333334" />
         <entry key="..\:/Users/cksgo/Desktop/UMC_Android/backup/cocktaildakk_android/app/src/main/res/layout/item_mainrec_cocktail.xml" value="0.20989583333333334" />
-        <entry key="..\:/Users/cksgo/Desktop/UMC_Android/backup/cocktaildakk_android/app/src/main/res/layout/item_recentsearch.xml" value="0.2115036231884058" />
+        <entry key="..\:/Users/cksgo/Desktop/UMC_Android/backup/cocktaildakk_android/app/src/main/res/layout/item_recentsearch.xml" value="0.75" />
         <entry key="..\:/Users/cksgo/Desktop/UMC_Android/backup/cocktaildakk_android/app/src/main/res/layout/item_searchtab_mainrec.xml" value="0.20989583333333334" />
+        <entry key="..\:/Users/cksgo/Desktop/UMC_Android/backup/cocktaildakk_android/app/src/main/res/menu/navi_menu.xml" value="0.20260416666666667" />
         <entry key="..\:/Users/cksgo/Desktop/UMC_Android/cocktail_dakk/cocktail_dakk/app/src/main/res/layout/fragment_home_detail.xml" value="0.20208333333333334" />
         <entry key="..\:/Users/cksgo/Desktop/UMC_Android/cocktail_dakk/cocktail_dakk/app/src/main/res/layout/fragment_keywordrecommand.xml" value="0.17603850050658562" />
         <entry key="..\:/Users/cksgo/Desktop/UMC_Android/cocktail_dakk/cocktail_dakk/app/src/main/res/layout/fragment_main.xml" value="0.17028985507246377" />
