+++ app/src/main/java/com/example/cocktail_dakk/ui/search/searchService/SearchService.kt
@@ -5,17 +5,49 @@ import com.example.cocktail_dakk.utils.getReposit
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
+import retrofit2.http.Query
+import kotlin.math.max
 
 class SearchService {
 
     private lateinit var searchView: SearchView
     private lateinit var pagingView: PagingView
+    private lateinit var filterView: FilterView
 
     fun setsearchView(searchView: SearchView) {
         this.searchView = searchView
     }
+
     fun setpagingView(pagingView: PagingView) {
         this.pagingView = pagingView
+    }
+
+    fun setfilterView(filterView: FilterView) {
+        this.filterView = filterView
+    }
+
+
+    fun filter(page : Int, keywordlist: List<String>,mindosu: Int, maxdosu: Int, drinklist: List<String>) {
+        val filterService = getReposit().create(SearchRetrofitInterface::class.java)
+        filterView.onFilterLoading()
+        filterService.filter(page,keywordlist,mindosu, maxdosu,drinklist).enqueue(object : Callback<SearchResponce>{
+            override fun onResponse(
+                call: Call<SearchResponce>,
+                response: Response<SearchResponce>
+            ) {
+                val resp = response.body()!!
+                Log.d("FilterView",resp.toString())
+                when (resp.code){
+                    1000 -> filterView.onFilterSuccess(resp.searchresult)
+                    else -> {
+                        filterView.onFilterFailure(resp.code,resp.message)
+                    }
+                }
+            }
+            override fun onFailure(call: Call<SearchResponce>, t: Throwable) {
+                searchView.onSearchFailure(400, "네트워크 오류 발생")
+            }
+        })
     }
 
     fun search(inputstr : String) {
