+++ app/src/main/res/layout/activity_splash.xml
@@ -19,18 +19,16 @@
             android:id="@+id/splash_logo_iv"
             android:layout_width="150dp"
             android:layout_height="168dp"
-            android:src="@drawable/splash_logo"
-            android:scaleType="fitXY"/>
-
-        <View
-            android:layout_width="10dp"
-            android:layout_height="10dp"/>
+            android:src="@mipmap/logo"
+            android:scaleType="centerCrop"
+            android:layout_marginBottom="10dp"
+            />
 
         <ImageView
             android:id="@+id/splash_title_iv"
-            android:layout_width="200dp"
-            android:layout_height="40dp"
-            android:src="@drawable/splash_title_temp"
+            android:layout_width="250dp"
+            android:layout_height="70dp"
+            android:src="@mipmap/cocktail_dakk"
             />
 
     </LinearLayout>
