+++ app/src/main/res/layout/item_recentsearch.xml
@@ -31,7 +31,7 @@
             android:layout_height="12dp"
             android:id="@+id/item_recentsearch_del_iv"
             android:layout_gravity="center"
-            android:src="@mipmap/icon_filter_close"
+            android:src="@drawable/search_history_delete"
             android:layout_marginEnd="10dp"
             />
     </LinearLayout>
